perf: default to 2 frames for calibration, remove dead code
Change default num_frames from 15 to 2. With 5%/95% skip, this
samples at ~30% and ~70% of the video - optimal for single-take
football footage. Reduces calibration time from 30s to 12s.

2 frames provides sufficient spatial diversity for all tested cameras
(GoPro, DJI, XTU) while being fast enough for interactive use.
Users can increase with --frames if needed.

Also remove unused calculate_step() (replaced by calculate_step_buffered).

+++ crates/reco-calibrate/src/akaze/mod.rs
@@ -138,10 +138,7 @@ impl Akaze {
         );
         trace!("Computing contrast factor finished.");
         // Pre-allocate diffusion buffers (reused across all evolution steps)
-        let (init_h, init_w) = (
-            evolutions[0].lt.height() as usize,
-            evolutions[0].lt.width() as usize,
-        );
+        let (init_h, init_w) = (evolutions[0].lt.height(), evolutions[0].lt.width());
         let mut diffusion_bufs = nonlinear_diffusion::DiffusionBuffers::new(init_h, init_w);
 
         for i in 1..evolutions.len() {
